% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mits.R
\name{read_mits}
\alias{read_mits}
\title{Download and import tidy data from the Melbourne Institute Time Series
collection}
\usage{
read_mits(
  tables = c("unemp", "wages"),
  download_dir = tempdir(),
  username = Sys.getenv("R_READMITS_USERNAME"),
  password = Sys.getenv("R_READMITS_PASSWORD"),
  port = NULL
)
}
\arguments{
\item{tables}{Vector of table(s) to download from the MITS site; currently
supported options are "unemp" (Unemployment expectations table) and
"wages" (Wages growth table)}

\item{download_dir}{Path to directory to which .xlsx file(s)
should be downloaded}

\item{username}{Your username for the Melbourne Institute Subscriber
Services website. See @details.}

\item{password}{Your password for the Melbourne Institute Subscriber
Services website. See @details.}

\item{port}{Port to use to connect to the Melbourne Institute website. If
left \code{NULL} (the default), an empty port will be identified
using \code{netstat::free_port()}.}
}
\value{
A tidy tbl_df
}
\description{
\code{read_mits()} establishes a headless browser session using \code{RSelenium}. Note
that this requires that Firefox is installed on your system.
}
\details{
By default, \code{read_mits()} will look for your username and password stored
as environmental variables. You can set these in an individual R session
\verb{Sys.setenv("R_READMITS_USERNAME" = <your-username>)} and
\verb{Sys.setenv("R_READMITS_PASSWORD" = <your-password>)}.

You can set these
variables in a persistent way (across multiple R sessions) by editing your
\code{.Renviron} file. The \code{usethis::edit_r_environ()} function provides a
convient way of opening this file. Add the following lines to this file:

\verb{R_READMITS_USERNAME=<your-username>}
\verb{R_READMITS_PASSWORD=<your-password>}

You can also supply the username and password directly when
calling \code{read_mits()}.
}
\examples{
\donttest{read_mits()}

}
